refactor: consolidate FiniteSet code into a single module
Move the Sort::finite_set*/Dynamic::as_finite_set impls out of sort.rs
and dynamic.rs into ast/finite_set.rs, and the finite_set tests out of
tests/lib.rs into a dedicated tests/finite_set.rs gated once via
#![cfg(z3_5_0_0)] at the file level, instead of scattering the cfg
attribute across every item and file.
